requirements.txt
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.10
google-api-core==2.4.0
google-auth==2.3.3
google-cloud-automl==2.5.2
googleapis-common-protos==1.54.0
grpcio==1.43.0
grpcio-status==1.43.0
ibm-cloud-sdk-core==3.14.0
ibm-watson==5.3.0
idna==3.3
install==1.3.5
praw==7.5.0
prawcore==2.3.0
proto-plus==1.19.8
protobuf==3.19.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.3.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
rsa==4.8
six==1.16.0
twilio==7.5.0
update-checker==0.18.0
urllib3==1.26.8
websocket-client==1.1.0
